# webApi/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample kanban application server. <br> Anass Lakhloufi - Anis Kharchi - Abdelmalek Benamirouche - Mohammed Naim Guerroui",
    "version": "1.0.0",
    "title": "Swagger Kanban",
    "termsOfService": "https://git-etudiants.lacl.fr/restProjet2019/anass-lakhloufi-anis-kharchi-abdelmalek-benamirouche-mohammed-naim-guerroui",
    "contact": {
      "email": "abdelmalek.benamirouche@etu.u-pec.fr"
    }
  },
  "host": "127.0.0.1:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/projects/{projectId}/cards": {
      "get": {
        "tags": [
          "card"
        ],
        "summary": "Find cards of a project by ID",
        "description": "Returns all cards of a single project",
        "operationId": "getCardsByProject",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "Id of project to add a card",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of Card objects",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Card"
              }
            }
          },
          "401": {
            "description": "Token value error"
          },
          "403": {
            "description": "Not a project member"
          },
          "500": {
            "description": "Server problem"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "tags": [
          "card"
        ],
        "summary": "Add a new Card to the project",
        "description": "",
        "operationId": "addCard",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "Id of project to add a card",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Card object needed to add into the project",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Card"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Card added with success"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Token value error"
          },
          "403": {
            "description": "Not a project member"
          },
          "500": {
            "description": "Server problem"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "JWT": {
      "description": "the token returned from the login",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "required": [
        "email",
        "password",
        "firstName",
        "lastName",
        "tel",
        "profession"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "tel": {
          "type": "string"
        },
        "profession": {
          "type": "string"
        },
        "image": {
          "type": "string"
        }
      }
    },
    "Card": {
      "type": "object",
      "required": [
        "title",
        "status",
        "type",
        "description"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string",
          "example": "project description"
        },
        "members": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        },
        "status": {
          "type": "string",
          "example": "Backlog"
        },
        "type": {
          "type": "string",
          "example": "Dev"
        }
      }
    },
...
  }
}